pyproject.toml
[project]
name = "pdftomd-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "docling>=2.41.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "marker-pdf>=1.8.1",
    "markitdown[pdf]>=0.1.2",
    "pymupdf4llm>=0.0.26",
    "pypdf2>=3.0.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]
